requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
colorama==0.4.4
Flask==1.1.2
Flask-Cors==3.0.9
gcloud==0.18.3
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
jws==0.1.3
MarkupSafe==1.1.1
oauth2client==4.1.2
packaging==20.7
pluggy==0.13.1
protobuf==3.14.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.9.9
pyparsing==2.4.7
Pyrebase4==4.4.1
pytest==6.1.2
python-jwt==2.0.1
requests==2.25.0
requests-toolbelt==0.9.1
rsa==4.6
six==1.15.0
toml==0.10.2
urllib3==1.26.2
Werkzeug==1.0.1
